new: Config.remoteServer && Pipeline.remoteServer() new: Config.apiRoot && Pipeline.apiRoot()
Now there is a single global location for the "standard" remote server configuration,
but every pipeline can override it individually.

Since it is an inheritable property, I think it would work to cluster
a bunch of pipelines together with a thin super-class to 'share' the common
remoteServer and apiRoot overrides. If that suites you :).